Fix quotes in password
This led to a bit of trouble because the database script didn't create
the table as expected (it failed to log in because the quotes in the
password were not escaped in the shell script).

I was unable to rebuild Docker with the new password quickly enough, so
I added a troubleshooting section in case this happens to you.